+++ packages/dds/tree/src/test/forestTestSuite.ts
@@ -39,6 +39,7 @@ import {
 	SchemaBuilder,
 	leaf,
 } from "../domains";
+import { typeboxValidator } from "../external-utilities";
 import { IdAllocator, JsonCompatible, brand, idAllocatorFromMaxId, mapIterable } from "../util";
 import {
 	FieldKinds,
@@ -419,7 +420,7 @@ export function testForest(config: ForestTestConfiguration): void {
 			const detachedFieldIndex = new DetachedFieldIndex(
 				"test",
 				idAllocatorFromMaxId() as IdAllocator<ForestRootId>,
-				undefined,
+				{ jsonValidator: typeboxValidator },
 				new RevisionTagCodec(),
 			);
 			const delta: DeltaFieldMap = new Map<FieldKey, DeltaFieldChanges>([
